changed value type in album, quote, and author class

+++ src/main/java/Author.java
@@ -1,20 +1,20 @@
 import java.io.Serializable;
 
 public class Author implements Serializable {
-    private int id;
+    private long id;
     private String firstName;
     private String lastName;
 
     public Author() {
     }
 
-    public Author(int id, String firstName, String lastName) {
+    public Author(long id, String firstName, String lastName) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
