+++ src/main/java/Quote.java
@@ -1,24 +1,24 @@
 import java.io.Serializable;
 
 public class Quote implements Serializable {
-    private int id;
+    private long id;
     private String content;
     private Author author;
 
     public Quote() {
     }
 
-    public Quote(int id, String content, Author author) {
+    public Quote(long id, String content, Author author) {
         this.id = id;
         this.content = content;
         this.author = author;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
